checkout process will start

+++ function.inc.php
@@ -94,9 +94,9 @@
     function removeBakeryFromCart($id){
         if(isset($_SESSION['FOOD_USER_ID'])){
             global $con;
-            $res=mysqli_query($con,"delete from bakery_cart where bakery_id='$id'");
+            $res=mysqli_query($con,"delete from bakery_cart where bakery_id='$id' and user_id=".$_SESSION['FOOD_USER_ID']);
         }else{
-          unset($_SESSION['cart'][$attr]);  
+          unset($_SESSION['cart'][$id]);  
         }
         
     }
